Rename `FileWrapper` -> `SourceLines`
Also move to `base_elements` module
